requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.0.0
django-cors-headers>=4.0.0
django-stubs>=5.0.0
djangorestframework-stubs>=3.15.0
drf-yasg>=1.21.0
PyYAML>=6.0
openpyxl>=3.1.0
reportlab>=4.0.0
